chore: move tagss context into an independent file
