+++ src/dataExplorer/components/Schema.tsx
@@ -18,6 +18,7 @@ import QueryProvider from 'src/shared/contexts/query'
 import {BucketProvider} from 'src/shared/contexts/buckets'
 import {MeasurementProvider} from 'src/dataExplorer/context/measurements'
 import {FieldsProvider} from 'src/dataExplorer/context/fields'
+import {TagsProvider} from 'src/dataExplorer/context/tags'
 
 // Types
 import {QueryScope} from 'src/types'
@@ -70,21 +71,23 @@ const Schema: FC = () => {
     <QueryProvider>
       <MeasurementProvider scope={scope}>
         <FieldsProvider scope={scope}>
-          <NewDataExplorerProvider scope={scope}>
-            <BucketProvider scope={scope}>
-              <div className="scroll--container">
-                <DapperScrollbars>
-                  <div className="data-schema">
-                    <BucketSelector />
-                    <div className="container-side-bar">
-                      <MeasurementSelector />
-                      <FieldsTags />
+          <TagsProvider scope={scope}>
+            <NewDataExplorerProvider>
+              <BucketProvider scope={scope}>
+                <div className="scroll--container">
+                  <DapperScrollbars>
+                    <div className="data-schema">
+                      <BucketSelector />
+                      <div className="container-side-bar">
+                        <MeasurementSelector />
+                        <FieldsTags />
+                      </div>
                     </div>
-                  </div>
-                </DapperScrollbars>
-              </div>
-            </BucketProvider>
-          </NewDataExplorerProvider>
+                  </DapperScrollbars>
+                </div>
+              </BucketProvider>
+            </NewDataExplorerProvider>
+          </TagsProvider>
         </FieldsProvider>
       </MeasurementProvider>
     </QueryProvider>
